TIP-843: Move array converter to enrich
