+++ src/Pim/Component/Connector/ArrayConverter/StandardToFlat/Product/ValueConverter/AbstractValueConverter.php
@@ -3,7 +3,7 @@
 namespace Pim\Component\Connector\ArrayConverter\StandardToFlat\Product\ValueConverter;
 
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
-use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\AttributeColumnsResolver;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\AttributeColumnsResolver;
 
 /**
  * @author    Adrien Pétremann <adrien.petremann@akeneo.com>
