+++ src/Pim/Component/Connector/spec/ArrayConverter/FlatToStandard/Product/FieldConverterSpec.php
@@ -6,8 +6,8 @@ use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Repository\GroupTypeRepositoryInterface;
 use Pim\Component\Connector\ArrayConverter\FlatToStandard\ConvertedField;
 use Pim\Component\Connector\ArrayConverter\FlatToStandard\FieldConverterInterface;
-use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\AssociationColumnsResolver;
-use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\FieldSplitter;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\AssociationColumnsResolver;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\FieldSplitter;
 
 class FieldConverterSpec extends ObjectBehavior
 {
