+++ src/Pim/Component/Connector/spec/ArrayConverter/FlatToStandard/Product/ValueConverter/NumberConverterSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\Val
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
-use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\FieldSplitter;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\FieldSplitter;
 
 class NumberConverterSpec extends ObjectBehavior
 {
