+++ src/Pim/Component/Connector/spec/ArrayConverter/FlatToStandard/Product/ValueConverter/ValueConverterRegistrySpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\ValueConverter;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\ValueConverter\ValueConverterInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\ValueConverter\ValueConverterInterface;
 
 class ValueConverterRegistrySpec extends ObjectBehavior
 {
