+++ src/Pim/Component/Connector/spec/ArrayConverter/FlatToStandard/ProductAssociationSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Component\Connector\ArrayConverter\FlatToStandard;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard;
 
 class ProductAssociationSpec extends ObjectBehavior
 {
