+++ tests/legacy/features/Context/FixturesContext.php
@@ -2529,7 +2529,7 @@ class FixturesContext extends BaseFixturesContext
     }
 
     /**
-     * @return \Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\AttributeColumnInfoExtractor
+     * @return \Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\AttributeColumnInfoExtractor
      */
     protected function getFieldExtractor()
     {
