+++ tests/legacy/features/Pim/Behat/Context/Storage/ProductModelStorage.php
@@ -13,7 +13,7 @@ use Behat\Gherkin\Node\TableNode;
 use Behat\MinkExtension\Context\RawMinkContext;
 use PHPUnit\Framework\Assert;
 use Pim\Component\Catalog\Repository\ProductModelRepositoryInterface;
-use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\AttributeColumnInfoExtractor;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\FlatToStandard\AttributeColumnInfoExtractor;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class ProductModelStorage extends RawMinkContext
